requirements.txt
absl-py==0.9.0
asgiref==3.2.7
astor==0.8.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
Django==3.0.6
djangorestframework==3.11.0
gast==0.2.2
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
idna==2.9
importlib-metadata==1.6.0
joblib==0.15.1
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.2.2
nltk==3.5
numpy==1.18.4
oauthlib==3.1.0
opt-einsum==3.2.1
protobuf==3.12.0
PySocks==1.7.1
pytz==2020.1
PyYAML==5.3.1
regex==2020.5.14
requests==2.23.0
requests-oauthlib==1.3.0
scikit-learn==0.23.0
scipy==1.4.1
six==1.14.0
sklearn==0.0
sqlparse==0.3.1
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
textblob==0.15.3
threadpoolctl==2.0.0
tqdm==4.46.0
tweepy==3.8.0
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
